Refactor all  file type handlers to be dispatched as strategies through StaticFilesHandler
Make fileStrategies static

Test all strategies through static files handler
