+++ src/main/java/me/karakelley/http/handlers/staticfilestrategies/DeleteResourceStrategy.java
@@ -1,4 +1,4 @@
-package me.karakelley.http.handlers;
+package me.karakelley.http.handlers.staticfilestrategies;
 
 import me.karakelley.http.http.Request;
 import me.karakelley.http.http.Response;
@@ -6,10 +6,10 @@ import me.karakelley.http.filesystem.PublicDirectory;
 import me.karakelley.http.http.responses.NoContent;
 import me.karakelley.http.server.Handler;
 
-public class DeleteResourceHandler implements Handler {
+public class DeleteResourceStrategy implements Handler {
   private PublicDirectory publicDirectory;
 
-  public DeleteResourceHandler(PublicDirectory publicDirectory) {
+  public DeleteResourceStrategy(PublicDirectory publicDirectory) {
     this.publicDirectory = publicDirectory;
   }
